PHP code review (PHPStan/Rector/Psalm) Adjust Psalm config (WIP)

+++ plugins/importExport/src/ModuleImportDc1.php
@@ -169,7 +169,7 @@ class ModuleImportDc1 extends Module
                 if ($this->importPosts($percent) === -1) {
                     Http::redirect($this->getURL() . '&do=ok');
                 } else {
-                    echo $this->progressBar(ceil($percent * 0.93) + 7);
+                    echo $this->progressBar((int) ceil((float) $percent * 0.93) + 7);
                 }
 
                 break;
